Separated python in each model file
