+++ mgmtsystem_nonconformity/__init__.py
@@ -1,3 +1,3 @@
 # -*- encoding: utf-8 -*-
 
-from . import mgmtsystem_nonconformity
+from . import models
