+++ mgmtsystem_nonconformity/models/mgmtsystem_nonconformity_origin.py
@@ -1,0 +1,66 @@
+# -*- encoding: utf-8 -*-
+##############################################################################
+#
+#    OpenERP, Open Source Management Solution
+#    Copyright (C) 2010 Savoir-faire Linux (<http://www.savoirfairelinux.com>).
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from openerp.osv import fields, orm
+
+
+class mgmtsystem_nonconformity_origin(orm.Model):
+    """
+    Origin of nonconformity of the management system
+    """
+    _name = "mgmtsystem.nonconformity.origin"
+    _description = "Origin of nonconformity of the management system"
+    _order = 'parent_id, sequence'
+
+    _columns = {
+        'id': fields.integer('ID', readonly=True),
+        'name': fields.char('Origin', size=50, required=True, translate=True),
+        'description': fields.text('Description'),
+        'sequence': fields.integer(
+            'Sequence',
+            help="Defines the order to present items",
+        ),
+        'parent_id': fields.many2one(
+            'mgmtsystem.nonconformity.origin',
+            'Group',
+        ),
+        'child_ids': fields.one2many(
+            'mgmtsystem.nonconformity.origin',
+            'parent_id',
+            'Childs',
+        ),
+        'ref_code': fields.char('Reference Code', size=20),
+    }
+
+    def name_get(self, cr, uid, ids, context=None):
+        ids = ids or []
+        reads = self.read(cr, uid, ids, ['name', 'parent_id'], context=context)
+        res = []
+        for record in reads:
+            name = record['name']
+            if record['parent_id']:
+                name = record['parent_id'][1] + ' / ' + name
+            res.append((record['id'], name))
+        return res
+
+    def _name_get_fnc(self, cr, uid, ids, prop, unknow_none, context=None):
+        res = self.name_get(cr, uid, ids, context=context)
+        return dict(res)
